% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_rgb_image.R
\name{add_rgb_image}
\alias{add_rgb_image}
\title{Add RGB product}
\usage{
add_rgb_image(s2_bands, session)
}
\arguments{
\item{s2_bands}{2-length list (one for TOA, one for BOA),
each element being a list of S2 bands, as defined in [s2_gui].}

\item{session}{Session}
}
\description{
Modal dialog to define an RGB image.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}, Pascal Obstetar (2020) \email{pascal.obstetar@gamil.com}
}
