% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extent.R
\name{base_map}
\alias{base_map}
\title{Create base_map}
\usage{
base_map(rv, map, session)
}
\arguments{
\item{rv}{Reactive value}

\item{map}{Map}

\item{session}{Session}
}
\value{
Base map
}
\description{
Create base_map
}
