% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_param.R
\name{check_param}
\alias{check_param}
\title{check_param}
\usage{
check_param(param_list, rv)
}
\arguments{
\item{param_list}{List param}

\item{rv}{Reactive values}
}
\value{
True or False
}
\description{
check_param
}
