% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_param.R
\name{check_param_list}
\alias{check_param_list}
\title{Check a parameter list}
\usage{
check_param_list(pm, type = "string", correct = TRUE)
}
\arguments{
\item{pm}{List of parameters or path of a JSON parameter file.}

\item{type}{Type of the output (see [print_message] for details).}

\item{correct}{Logical: if TRUE (default), the function corrects
some incoherences (e.g. timewindow of length 1 is transformed in length 2)
and returns the corrected list as output; if false, only checking is
performed, and the output is NULL if no errors occur.}
}
\value{
In case of errors, depending on `type` argument, output can be
 a vector of errors (if `type = 'string'`),
 the first error occurred (if `type = 'error'`)
 or a set of warnings (if `type = 'warning'`).
 If no errors occur, output is the corrected parameter list if
 `correct = TRUE` or NULL otherwise.
}
\description{
Check that the parameter list (or JSON parameter file)
 is in the correct format, and then speficied values are coherent with
 parameters.
}
\note{
License: GPL 3.0
}
\author{
Pascal Obstetar (2020) \email{pascal.obstetar@onf.fr}
}
