% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getData_WorldClim.R
\name{getDataWorldClim}
\alias{getDataWorldClim}
\title{getDataWorldClim}
\usage{
getDataWorldClim(
  name = "worldclim",
  download = TRUE,
  path = "~",
  period = "1970-2000",
  gcm = "BCC-CSM2-MR",
  ssp = 126,
  progress_bar = TRUE,
  ...
)
}
\arguments{
\item{name}{}

\item{download}{}

\item{path}{}

\item{period}{}

\item{gcm}{}

\item{ssp}{}

\item{...}{}
}
\value{

}
\description{
getDataWorldClim
}
