% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_param.R
\name{import_param_list}
\alias{import_param_list}
\title{import_param_list}
\usage{
import_param_list(session, rv)
}
\arguments{
\item{session}{Shiny session}

\item{rv}{Reactivevalue list}
}
\value{
A param list
}
\description{
import_param_list
}
