% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_list_indices.R
\name{list_indices}
\alias{list_indices}
\title{List spectral indices}
\usage{
list_indices(values, pattern = "", all = FALSE)
}
\arguments{
\item{values}{A vector of attributes which will be returned, being
one or more within the followings:
- `n_index`: internal index identifiers;
- `name`: index name;
- `longname`: index description;
- `link`: URL to the index description page;
- `s2_formula`: expression containing the formula to compute the index;
- `s2_formula_mathml`: MathML version of the formula;
- `checked`: logical (TRUE for verified indices);
- `a`, `b`, `x`: parameter values (NA for non required parameters).}

\item{pattern}{A regular expression on index names.}

\item{all}{Logical: if TRUE, all the indices retrieved from
[IDB](http://www.indexdatabase.de/db/is.php?sensor_id=96) are returned;
if FALSE (default), only indices checked by the authors are returned.}
}
\value{
A data.frame with the required information. The table contains
 also the following attributes:
 - `creation_date`: timestamp of the creation date of the indices archive;
 - `pkg_version`: version of the `sen2r` package used to
     create the indices archive.
}
\description{
Return a table with attributes of the spectral indices
 computable with the package.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
