% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extent.R
\name{load_extent_draw}
\alias{load_extent_draw}
\title{load_extent_draw}
\usage{
load_extent_draw(extent_ns_name, rv, session)
}
\arguments{
\item{extent_ns_name}{Name of the namespace to be used}

\item{rv}{Reactive values}

\item{session}{Session}
}
\description{
load_extent_draw
}
