% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extent.R
\name{load_extent_gpkg}
\alias{load_extent_gpkg}
\title{load_extent_gpkg}
\usage{
load_extent_gpkg(rv, session)
}
\arguments{
\item{rv}{Reactive values}

\item{session}{Session}
}
\description{
load_extent_gpkg
}
