% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project.R
\name{mod_project_ui}
\alias{mod_project_ui}
\alias{mod_project_server}
\title{mod_project_ui and mod_project_server}
\usage{
mod_project_ui(id)

mod_project_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
