% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nn.R
\name{nn}
\alias{nn}
\title{Replace NULL with character()}
\usage{
nn(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
`character(0)` if `x==NULL`, `x` elsewhere
}
\description{
Internal function: return character(0) instead of NULL.
 This is sometimes needed not to return error when applying some functions.
}
\note{
License: GPL 3.0
}
\examples{
tryCatch(basename(NULL), error = print) # error
basename(character()) # ok
basename(shinyCNES:::nn(NULL)) # ok
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
