% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_theia_login.R
\name{read_theia_login}
\alias{read_theia_login}
\title{read_theia_login}
\usage{
read_theia_login(apitheia_path = NA)
}
\arguments{
\item{apitheia_path}{Path api THEIA}
}
\description{
read_theia_login
}
