% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_indice.R
\name{s2_indice}
\alias{s2_indice}
\title{s2_indice}
\usage{
s2_indice(
  path_data = "~",
  path_indice = "~",
  project_name,
  collection = "Sentinel2_L2A_THEIA",
  tiles = NULL,
  indices = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{Path data}

\item{path_indice}{Path indice}

\item{project_name}{Project name}

\item{collection}{Collection}

\item{tiles}{Tiles}

\item{indices}{Indices}

\item{progress_bar}{Progress bar if shiny run}
}
\value{

}
\description{
s2_indice
}
