% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_orbits.R
\name{s2_orbits}
\alias{s2_orbits}
\title{Load Sentinel-2 orbits}
\usage{
s2_orbits()
}
\value{
An sf spatial object containing the extent of the orbits.
}
\description{
Load the vector object of the Sentinel-2 orbits.
}
\note{
License: GPL 3.0
}
\author{
Pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
