% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_preview_img.R
\name{s2_preview_img}
\alias{s2_preview_img}
\title{s2_preview_img}
\usage{
s2_preview_img(
  path_data = "~",
  project_name,
  collection = "Sentinel2_L2A_THEIA",
  tile = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{}
}
\value{

}
\description{
s2_preview_img
}
