% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_preview_indice.R
\name{s2_preview_indice}
\alias{s2_preview_indice}
\title{s2_preview_indice}
\usage{
s2_preview_indice(
  path_data = "~",
  project_name,
  collection = "Sentinel2_L2A_THEIA",
  tile = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{Path data}

\item{project_name}{Project name}

\item{collection}{Collection}

\item{tile}{Tile}

\item{progress_bar}{Progress bar if shiny run}
}
\value{

}
\description{
s2_preview_indice
}
