% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_preview_rgb.R
\name{s2_preview_rgb}
\alias{s2_preview_rgb}
\title{s2_preview_rgb}
\usage{
s2_preview_rgb(
  path_data = "~",
  project_name,
  collection = "Sentinel2_L2A_THEIA",
  tile = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{Path data}
}
\value{

}
\description{
s2_preview_rgb
}
