% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_preview_srtm.R
\name{s2_preview_srtm}
\alias{s2_preview_srtm}
\title{s2_preview_srtm}
\usage{
s2_preview_srtm(
  path_data = "~",
  project_name,
  path_tif = "~",
  tile = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{Path data}

\item{project_name}{project name}

\item{tile}{Tile}

\item{progress_bar}{Progress bar if shiny run}

\item{path}{Path}
}
\value{

}
\description{
s2_preview_srtm
}
