% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_rgb.R
\name{s2_rgb}
\alias{s2_rgb}
\title{s2_preview}
\usage{
s2_rgb(
  path_data = "~",
  path_rgb = "~",
  project_name,
  collection = "Sentinel2_L2A_THEIA",
  tiles = NULL,
  rgbs = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{path_data}{Path data}
}
\value{

}
\description{
s2_preview
}
