% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_s2_tiles.R
\name{s2_tiles}
\alias{s2_tiles}
\title{Load Sentinel-2 tiles}
\usage{
s2_tiles()
}
\value{
An sf spatial object containing the extent of the tiles.
}
\description{
Load the vector object of the Sentinel-2 tiles.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}, pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
