% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_srtm.R, R/fct_worldclim.R
\name{srtm}
\alias{srtm}
\title{srtm}
\usage{
srtm(
  session,
  path_project,
  path_tif,
  srtm_multiple = "9,15,33",
  time_interval = NULL,
  spatial_extent = NULL,
  progress_bar = TRUE
)

srtm(
  session,
  path_project,
  path_tif,
  srtm_multiple = "9,15,33",
  time_interval = NULL,
  spatial_extent = NULL,
  progress_bar = TRUE
)
}
\arguments{
\item{session}{Session}

\item{path_project}{Path project}

\item{path_tif}{Path tif}

\item{srtm_multiple}{Multiple SRTM}

\item{time_interval}{Time interval}

\item{spatial_extent}{Spatial extent}

\item{progress_bar}{Progress bar if shiny run}

\item{list_indice}{List of derivated indices. Default vector from 'exposition' and 'ombrage'}
}
\value{



}
\description{
srtm

srtm
}
