% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_theia_login.R
\name{theia_modal}
\alias{theia_modal}
\title{theia_modal}
\usage{
theia_modal(username = NA, password = NA, session)
}
\arguments{
\item{username}{Theia login}

\item{password}{Theia password}
}
\description{
theia_modal
}
