% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tiles_intersects.R
\name{tiles_intersects}
\alias{tiles_intersects}
\title{Select the tiles intersecting the extent}
\usage{
tiles_intersects(extent, all = FALSE, out_format = "id", .s2tiles = NULL)
}
\arguments{
\item{extent}{`sf` object with the spatial extent.}

\item{all}{logical: if TRUE, all the tiles overlapping the extent are
provided;
if FALSE (default), unnecessary tiles are skipped.
Unnecessary tiles are tiles which overlaps the extent for an area already
covered by another tile.
In case the extent is all included in an overlapping area, only one of the
two candidate tiles is returned (the first in alphabetical order).}

\item{out_format}{character: if "sf", the spatial object of the overlapping tiles
is returned; if "id" (default), a character vector with the tile IDs.}

\item{.s2tiles}{output of [s2_tiles()] function (it is possible to pass it
in order to speed up the execution;
otherwise leave to NULL and it will be generated within the function).}
}
\value{
the tiles intersecting the extent (see argument `out_format`).
}
\description{
Function which returns the tile IDs of the Sentinel-2 tiles
 which overlap a provided extent.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
