% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extent.R
\name{update_extent}
\alias{update_extent}
\title{update_extent}
\usage{
update_extent(
  rv,
  extent_source,
  custom_source = NA,
  map = "view_map_extent",
  session
)
}
\arguments{
\item{rv}{Reactive value}

\item{extent_source}{Extent source}

\item{custom_source}{Custom source}

\item{map}{Map}

\item{session}{Session}
}
\value{
Update the map
}
\description{
update_extent
}
