% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_theia_login.R
\name{write_theia_login}
\alias{write_theia_login}
\title{write_theia_login}
\usage{
write_theia_login(username, password, apitheia_path = NA)
}
\arguments{
\item{username}{Theia login}

\item{password}{Theia password}

\item{apitheia_path}{Path theia}
}
\description{
write_theia_login
}
